fix: address inbox limit review feedback
